A project using J2EE with Oracle 10g
This project is mainly undertaken keeping in mind the growing trend of
online ordering systems. The idea is to bring the delicious foods online
in an innovative way. This system has a registration system where
customers can fill in their details once, and when they revisit to order
food, they need not provide them again, the details stored will be
retrieved from the database. The system also maintains a database for
available products for sale and the details of ordering. And finally
after ordering, customers are provided with a bill and the estimated
time of delivery.